+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - Fixed a bug in `--{_:()=>null}._` which was rejected because tail parsing was receiving the wrong set of parameters for the update.
 - Fixes a bug in `let instanceof x` where the code was checking `in` twice, where it should check `in` and `instanceof` instead.
 - Fixed an edge case in annexB / webcompat handling of function statement names.
+- Fixed an issue where generator / async function names were treated as var rather than lex
 
 ## v2.0.1